AppImage: include libOpenGL.so.0 fallback
Currently Ubuntu 25.10 live DVD (ubuntu-25.10-desktop-amd64.iso)
does not include the library, preventing the GUI from being
run.  On the other hand, if used unconditionally, it causes
the warning (even in the U25.10) for which it was removed:
<https://github.com/linuxdeploy/linuxdeploy/issues/152>

So preload the library if is in the system and use the bundled just as
a fallback.

+++ data/scripts/Linux-AppImage/create-appimage.sh
@@ -149,6 +149,10 @@ for n in $EXCLUDE_LIST; do
         if [ "$n" = libjack.so.0 ] || [ "$n" = libpipewire-0.3.so.0 ]; then
                 continue
         fi
+        # not included in U25.10 live
+        if [ "$n" = libOpenGL.so.0 ]; then
+                continue
+        fi
         if [ -f "$APPPREFIX/lib/$n" ]; then
                 rm "$APPPREFIX/lib/$n"
         fi
